chore: improve Input exports
